+++ resources/ts/Components/Layout/HeadPageLayout.vue
@@ -1,0 +1,54 @@
+<!--
+    Exper-Dat-Reader is a system to read encrypted .dat files and dump their data into .done.dat files.
+    Copyright (C) 2022  Mestre-Tramador
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <https://www.gnu.org/licenses/>.
+-->
+
+<template>
+    <main-layout>
+        <div class="flex">
+            <slot name="icon"></slot>
+
+            <div class="mx-2 my-auto text-lg text-blue-400 cursor-default">
+                {{ title }}
+            </div>
+        </div>
+
+        <hr />
+
+        <slot name="page"></slot>
+    </main-layout>
+</template>
+
+<script lang="ts">
+import { defineComponent } from "vue";
+
+import MainLayout from "./MainLayout.vue";
+
+export default defineComponent({
+    components: {
+        MainLayout
+    },
+    props: {
+        /**
+         * The Title of the page.
+         */
+        title: {
+            type: String,
+            default: ""
+        }
+    }
+});
+</script>