+++ docs/CHANGELOG.md
@@ -10,6 +10,23 @@ Versioning].
 - Changelog on different languages
 - GitHub Pull Requests templates
 
+## [0.0.7] - 2023-02-13
+
+### Created
+
+- Dashboard Vue Page
+- Page Vue Layout
+
+### Implemented
+
+- Last Done Dat API Route
+- Typescript Response Data interfaces with keys
+
+### Improved
+
+- Versions of Composer and NPM packages
+- PHP files commentaries and namespace usages
+
 ## [0.0.6] - 2022-10-18
 
 ### Created
@@ -179,7 +196,8 @@ Versioning].
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
-[Unreleased]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.6...HEAD
+[Unreleased]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.7...HEAD
+[0.0.7]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.5...v0.0.6
 [0.0.5]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/Mestre-Tramador/Exper-Dat-Reader/compare/v0.0.3...v0.0.4